+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### Added
 
-- Type for `radius`
+- Typings for `radius`, `if`, `bgWhite`
 - `left`, `right`, `top`, `bottom` props
+- `z` alias prop for `zIndex`
 
 ## [1.2.0] - 2019-12-13